BLUGA: Fix warnings

+++ libs/B-luga-graphics/include/B-luga-graphics/SpriteSystems.hpp
@@ -127,7 +127,7 @@ namespace Systems {
                         Maths::intToFloatConservingDecimals(position.y));
                     size = calculateSize(sprite);
                     sprite.drawPro(
-                        Raylib::Rectangle(0.F, 0.F, sprite.getTextureWidth(), sprite.getTextureHeight()),
+                        Raylib::Rectangle(0.F, 0.F, static_cast<float>(sprite.getTextureWidth()), static_cast<float>(sprite.getTextureHeight())),
                         Raylib::Rectangle(spritePos.x, spritePos.y, size.x, size.y),
                         Raylib::Vector2(origin.x, origin.y),
                         rotation,
